Implement home screen service content
